+++ src/parser/inline.test.ts
@@ -90,13 +90,13 @@ describe('Unit: parser/inline', () => {
     });
 
     it('hashtag', () => {
-      assert.deepStrictEqual(inspect(parser('#a#')), [['#', 'a#'], '']);
-      assert.deepStrictEqual(inspect(parser('#a#b')), [['#', 'a#', 'b'], '']);
+      assert.deepStrictEqual(inspect(parser('#a#')), [['#a#'], '']);
+      assert.deepStrictEqual(inspect(parser('#a#b')), [['#a#b'], '']);
       assert.deepStrictEqual(inspect(parser('#a')), [['<a class="hashtag" rel="noopener">#a</a>'], '']);
       assert.deepStrictEqual(inspect(parser('#a\nb\n#c\n[#d]')), [['<a class="hashtag" rel="noopener">#a</a>', '<br>', 'b', '<br>', '<a class="hashtag" rel="noopener">#c</a>', '<br>', '<a href="#index:d" rel="noopener">d</a>'], '']);
-      assert.deepStrictEqual(inspect(parser('##a')), [['##', 'a'], '']);
-      assert.deepStrictEqual(inspect(parser('a#b')), [['a#', 'b'], '']);
-      assert.deepStrictEqual(inspect(parser('あ#b')), [['あ#', 'b'], '']);
+      assert.deepStrictEqual(inspect(parser('##a')), [['##a'], '']);
+      assert.deepStrictEqual(inspect(parser('a#b')), [['a#b'], '']);
+      assert.deepStrictEqual(inspect(parser('あ#b')), [['あ#b'], '']);
       assert.deepStrictEqual(inspect(parser('a\n#b')), [['a', '<br>', '<a class="hashtag" rel="noopener">#b</a>'], '']);
       assert.deepStrictEqual(inspect(parser('a\\\n#b')), [['a', '<span class="linebreak"> </span>', '<a class="hashtag" rel="noopener">#b</a>'], '']);
       assert.deepStrictEqual(inspect(parser('*a*#b')), [['<em>a</em>', '<a class="hashtag" rel="noopener">#b</a>'], '']);