+++ src/parser/inline/autolink/account.test.ts
@@ -1,24 +1,26 @@
-import { account } from './account';
+import { autolink } from '../autolink';
 import { some } from '../../../combinator';
 import { inspect } from '../../../debug.test';
 
 describe('Unit: parser/inline/autolink/account', () => {
   describe('account', () => {
-    const parser = some(account);
+    const parser = some(autolink);
 
     it('invalid', () => {
       assert.deepStrictEqual(inspect(parser('')), undefined);
       assert.deepStrictEqual(inspect(parser('@')), [['@'], '']);
-      assert.deepStrictEqual(inspect(parser('@_')), [['@_'], '']);
-      assert.deepStrictEqual(inspect(parser('@-')), [['@-'], '']);
-      assert.deepStrictEqual(inspect(parser('@_a')), [['@_a'], '']);
-      assert.deepStrictEqual(inspect(parser('@-a')), [['@-a'], '']);
+      assert.deepStrictEqual(inspect(parser('@+')), [['@'], '+']);
+      assert.deepStrictEqual(inspect(parser('@_')), [['@'], '_']);
+      assert.deepStrictEqual(inspect(parser('@-')), [['@'], '-']);
+      assert.deepStrictEqual(inspect(parser('@.')), [['@'], '.']);
       assert.deepStrictEqual(inspect(parser('@a@')), [['@a@'], '']);
+      assert.deepStrictEqual(inspect(parser('@a@b')), [['@a@b'], '']);
+      assert.deepStrictEqual(inspect(parser('@a@b@c')), [['@a@b@c'], '']);
       assert.deepStrictEqual(inspect(parser('@ab@')), [['@ab@'], '']);
       assert.deepStrictEqual(inspect(parser('@a@b')), [['@a@b'], '']);
       assert.deepStrictEqual(inspect(parser('@@')), [['@@'], '']);
       assert.deepStrictEqual(inspect(parser('@@a')), [['@@a'], '']);
-      assert.deepStrictEqual(inspect(parser('a@b')), undefined);
+      assert.deepStrictEqual(inspect(parser('@@@a')), [['@@@a'], '']);
       assert.deepStrictEqual(inspect(parser(' @a')), undefined);
     });
 
@@ -33,8 +35,6 @@ describe('Unit: parser/inline/autolink/account', () => {
       assert.deepStrictEqual(inspect(parser('@a--b')), [['<a class="account" rel="noopener">@a</a>'], '--b']);
       assert.deepStrictEqual(inspect(parser('@http://host')), [['<a class="account" rel="noopener">@http</a>'], '://host']);
       assert.deepStrictEqual(inspect(parser('@ttp://host')), [['<a class="account" rel="noopener">@ttp</a>'], '://host']);
-      assert.deepStrictEqual(inspect(parser('@a#')), [['<a class="account" rel="noopener">@a</a>'], '#']);
-      assert.deepStrictEqual(inspect(parser('@a#1')), [['<a class="account" rel="noopener">@a</a>'], '#1']);
     });
 
   });