+++ src/parser/inline/autolink/account.ts
@@ -1,15 +1,7 @@
 import { AutolinkParser } from '../../inline';
-import { union, some, subline, focus, verify } from '../../../combinator';
-import { unescsource } from '../../source/unescapable';
-import { defrag } from '../../util';
+import { subline, focus } from '../../../combinator';
 import { html } from 'typed-dom';
 
-export const account: AutolinkParser.AccountParser = subline(union([
-  verify(focus(
-    /^@[a-zA-Z0-9]+(?:-[0-9a-zA-Z]+)*/,
-    source => [[html('a', { class: 'account', rel: 'noopener' }, source)], '']),
-    (_, rest) => !rest.startsWith('@')),
-  focus(
-    /^(?:@[a-zA-Z0-9.+_-]*)+/,
-    defrag(some(unescsource))),
-]));
+export const account: AutolinkParser.AccountParser = subline(focus(
+  /^@[a-zA-Z0-9]+(?:-[0-9a-zA-Z]+)*/,
+  source => [[html('a', { class: 'account', rel: 'noopener' }, source)], '']));