+++ src/parser/inline/autolink/channel.ts
@@ -1,14 +1,13 @@
 import { AutolinkParser } from '../../inline';
-import { sequence, some, subline, verify, fmap } from '../../../combinator';
+import { sequence, some, subline, fmap } from '../../../combinator';
 import { account } from './account';
-import { hashtag_ } from './hashtag';
+import { hashtag } from './hashtag';
 import { stringify } from '../../util';
 import { html } from 'typed-dom';
 
-export const channel: AutolinkParser.ChannelParser = subline(fmap(verify(
+export const channel: AutolinkParser.ChannelParser = subline(fmap(
   sequence([
-    verify(account, ([node]) => node instanceof HTMLAnchorElement),
-    some(verify(hashtag_, ([node]) => node instanceof HTMLAnchorElement)),
+    account,
+    some(hashtag),
   ]),
-  (_, rest) => !rest.startsWith('#')),
   ns => [html('a', { class: 'channel', rel: 'noopener' }, stringify(ns))]));